requirements.txt
# ========================================
# SISTEMA GIL - GESTIÓN INTELIGENTE DE LABORATORIOS
# Dependencias del proyecto
# Centro Minero de Sogamoso - SENA
# ========================================

# Base de datos
mysql-connector-python==8.2.0
PyMySQL==1.1.0

# Seguridad y autenticación
bcrypt==4.1.2
cryptography==42.0.5
PyJWT==2.8.0

# Framework web
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6

# APIs y servicios web
requests==2.31.0
urllib3==2.1.0

# Procesamiento de datos
pandas==2.1.4
numpy==1.26.4
openpyxl==3.1.2

# Inteligencia Artificial - Reconocimiento de Voz
speechrecognition==3.10.0
pyaudio==0.2.13
pydub==0.25.1
webrtcvad==2.0.10

# Inteligencia Artificial - Reconocimiento de Imágenes
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
tensorflow==2.15.0
keras==2.15.0
Pillow==10.1.0
scikit-image==0.22.0
imageio==2.33.1

# Modelos preentrenados
tensorflow-hub==0.15.0
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2

# Procesamiento de audio
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.4

# Generación de códigos QR
qrcode[pil]==7.4.2
python-barcode==0.15.1

# Validación y serialización
marshmallow==3.20.2
cerberus==1.3.5
jsonschema==4.20.0

# Utilidades de fecha y tiempo
python-dateutil==2.8.2
pytz==2023.3

# Logging y monitoreo
colorlog==6.8.0
python-json-logger==2.0.7

# Variables de entorno
python-dotenv==1.0.0

# Pruebas
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Documentación API
flask-restx==1.3.0
flasgger==0.9.7.1

# Procesamiento de archivos
python-magic==0.4.27
pathlib2==2.3.7

# Comunicación en tiempo real
python-socketio==5.10.0
eventlet==0.33.3

# Tareas en segundo plano
celery==5.3.4
redis==5.0.1

# Configuración y argumentos
configargparse==1.7
click==8.1.7

# Desarrollo y herramientas
black==23.12.0
flake8==6.1.0
autopep8==2.0.4